Added the Load Data Function.
Moved the load data function from the AI algorithm to the formatter. Created a new function that just removes labels that are off, and added some try blocks to catch some errors that occur when the labels being compared do not exist. May need to work out a continue from the list or see how the list can be updated as the loop goes on.