Started testing for repeating labels.
Created a function that created a list of labels when loading the data, helps observe whether the data has the proper labels and the frequency of said labels. 73% of the labels were unique values that were only found in one case scenario and should be excluded from the full list of labels.